Make clippy happy

+++ src/editors/mod.rs
@@ -39,9 +39,9 @@ pub fn raven_color_to_rgb(raven_color: u8) -> egui::Color32 {
     let r = raven_color & 0x03;
     let g = (raven_color & 0x0c) >> 2;
     let b = (raven_color & 0x30) >> 4;
-    let cr = ((r << 6) | (r << 4) | (r << 2) | r) as u8;
-    let cg = ((g << 6) | (g << 4) | (g << 2) | g) as u8;
-    let cb = ((b << 6) | (b << 4) | (b << 2) | b) as u8;
+    let cr = (r << 6) | (r << 4) | (r << 2) | r;
+    let cg = (g << 6) | (g << 4) | (g << 2) | g;
+    let cb = (b << 6) | (b << 4) | (b << 2) | b;
     egui::Color32::from_rgb(cr, cg, cb)
 }
